refactor: clean up the packages in the app-series
Move some items into a view package inside list, which hosts all of the items for the recycler view
