+++ app-series/src/main/java/com/chesire/nekome/app/series/list/DialogHandler.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.app.series
+package com.chesire.nekome.app.series.list
 
 import android.content.Context
 import androidx.lifecycle.LifecycleOwner
@@ -6,6 +6,7 @@ import com.afollestad.materialdialogs.MaterialDialog
 import com.afollestad.materialdialogs.lifecycle.lifecycleOwner
 import com.afollestad.materialdialogs.list.listItems
 import com.afollestad.materialdialogs.list.listItemsMultiChoice
+import com.chesire.nekome.app.series.SeriesPreferences
 import com.chesire.nekome.core.R
 import com.chesire.nekome.core.flags.SortOption
 import com.chesire.nekome.core.flags.UserSeriesStatus
